new-update

+++ src/app/api/posts/myneed/[id]/route.js
@@ -1,8 +1,31 @@
 import { NextResponse } from "next/server";
 import { getAuthSession } from "../../../../../utils/auth";
 
-export const DELETE = async (req, { params }) => {
-  const { id } = params;
+// export const DELETE = async (req, { params }) => {
+//   const { id } = params;
+
+//   const session = await getAuthSession();
+
+//   if (!session) {
+//     return new NextResponse(
+//       JSON.stringify({ message: "Not Authenticated!" }, { status: 401 })
+//     );
+//   }
+
+//   try {
+//     const data = await prisma.post.delete({
+//       where: {
+//         id: id,
+//       },
+//     });
+//     return new NextResponse(JSON.stringify(data), { status: 200 });
+//   } catch (error) {
+//     console.log(error);
+//   }
+// };
+
+export default async function handler(req, res) {
+  const { id } = req.query;
 
   const session = await getAuthSession();
 
@@ -12,14 +35,19 @@ export const DELETE = async (req, { params }) => {
     );
   }
 
-  try {
-    const data = await prisma.post.delete({
-      where: {
-        id: id,
-      },
-    });
-    return new NextResponse(JSON.stringify(data), { status: 200 });
-  } catch (error) {
-    console.log(error);
+  if (req.method === "DELETE") {
+    try {
+      await prisma.post.delete({
+        where: {
+          id: id,
+        },
+      });
+      return new NextResponse.stringify(
+        { message: "Post deleted successfully" },
+        { status: 200 }
+      );
+    } catch (error) {
+      console.log(error);
+    }
   }
-};
+}
